requirements.txt
celery==5.4.0
Django==5.1.4
django-ninja==1.3.0
django-redis==5.4.0
django-role-permissions==3.2.0
flower==2.0.1
python-dotenv==1.0.1
redis==3.5.3
requests==2.32.3